% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildForestedLandscape.R
\name{buildForestedLandscape}
\alias{buildForestedLandscape}
\title{Builds an object of SpatialPixelsLandscape for a target area}
\usage{
buildForestedLandscape(boundaries, grid, dataset_path = "~/OneDrive/Datasets/")
}
\arguments{
\item{boundaries}{spatial polygons delimiting the target area}

\item{grid}{raster definition to resample topography at the desired resolution}

\item{dataset_path}{path to the 'Datasets' directory}
}
\description{
Builds an object of SpatialPixelsLandscape for a target area
}
