% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSoils.R
\name{buildSoils}
\alias{buildSoils}
\title{Builds a list of soil objects for a set of target point locations}
\usage{
buildSoils(
  pts,
  dataset_path = "~/OneDrive/Datasets/",
  widths = c(300, 700, 1000, 2000)
)
}
\arguments{
\item{pts}{target points}

\item{dataset_path}{path to the 'Datasets' directory}

\item{widths}{soil layer widths (in mm)}
}
\description{
Builds a list of soil objects for a set of target point locations
}
