% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTopography.R
\name{buildTopography}
\alias{buildTopography}
\title{Extracts a raster with topography variables from DEM}
\usage{
buildTopography(boundaries, grid = NULL, dataset_path = "~/OneDrive/Datasets/")
}
\arguments{
\item{boundaries}{spatial polygons delimiting the target area}

\item{grid}{raster definition to resample topography at the desired resolution}

\item{dataset_path}{path to the 'Datasets' directory}
}
\description{
Extracts a raster with elevation, slope and aspect for a target area
}
