% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultInitializationParams.R
\name{defaultInitializationParams}
\alias{defaultInitializationParams}
\title{Default initialization params}
\usage{
defaultInitializationParams()
}
\description{
Creates a list of default parameters for landscape initialization procedures
}
