% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getForestList.R
\name{getForestList}
\alias{getForestList}
\title{Defines list of 'forest' objects}
\usage{
getForestList(
  pts,
  fib,
  lct,
  merge_trees = TRUE,
  correct_lidar = TRUE,
  dataset_path = "~/OneDrive/Datasets/"
)
}
\arguments{
\item{pts}{target locations}

\item{fib}{forest imputation basis, returned by function buildForestImputationBasis}

\item{lct}{vector of land cover type}

\item{merge_trees}{boolean flag to simplify trees into diameter classes}

\item{correct_lidar}{boolean flag to correct density and mean height using information derived from lidar}

\item{dataset_path}{path to the 'Datasets' directory}
}
\description{
Defines list of 'forest' objects for a set of target locations
}
