% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopography.R
\name{getTopography}
\alias{getTopography}
\title{Extracts a raster with topography variables from DEM}
\usage{
getTopography(
  boundaries,
  grid = NULL,
  mask = TRUE,
  dataset_path = "~/OneDrive/Datasets/"
)
}
\arguments{
\item{boundaries}{spatial object (normally polygons) delimiting the target area}

\item{grid}{raster definition to resample topography at the desired resolution}

\item{mask}{boolean flag to mask raster according to boundaries}

\item{dataset_path}{path to the 'Datasets' directory}
}
\description{
Extracts a raster with elevation, slope and aspect for a target area
}
