% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildForestedLandscape.R
\name{buildForestedLandscape}
\alias{buildForestedLandscape}
\title{Builds an object of SpatialPixelsLandscape for a target area}
\usage{
buildForestedLandscape(
  boundaries,
  grid,
  fib,
  initializationParams = defaultInitializationParams(),
  dataset_path = "~/OneDrive/Datasets/"
)
}
\arguments{
\item{boundaries}{spatial polygons delimiting the target area}

\item{grid}{raster definition to resample topography at the desired resolution}

\item{fib}{forest imputation basis, returned by function buildForestImputationBasis}

\item{initializationParams}{list with initialisation options}

\item{dataset_path}{path to the 'Datasets' directory}
}
\value{
an object of class 'SpatialPixelsLandscape'
}
\description{
This function makes internal calls to other package functions to obtain topography, land cover,
soil and forest data over the target area.
}
\seealso{
buildForestImputationBasis
}
