% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildForestImputationBasis.R
\name{buildForestImputationBasis}
\alias{buildForestImputationBasis}
\title{Assembles forest plot data for forest stand imputation}
\usage{
buildForestImputationBasis(dataset_path = "~/OneDrive/Datasets/")
}
\arguments{
\item{dataset_path}{path to the 'Datasets' directory}
}
\description{
Assembles forest plot data for forest stand imputation
}
