% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultInitializationParams.R
\name{defaultInitializationParams}
\alias{defaultInitializationParams}
\title{Default initialization params}
\usage{
defaultInitializationParams()
}
\value{
A list with the following elements:
 \itemize{
   \item{correct_lidar: boolean}
 }
}
\description{
Creates a list of default parameters for landscape initialization procedures
}
