% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSoilGridsParams.R
\name{getSoilGridsParams}
\alias{getSoilGridsParams}
\title{Builds a list of soil objects for a set of target point locations}
\usage{
getSoilGridsParams(
  pts,
  widths = c(300, 700, 1000, 2000),
  modify_soil_depth = TRUE,
  dataset_path = "~/OneDrive/Datasets/"
)
}
\arguments{
\item{pts}{target locations}

\item{widths}{soil layer widths (in mm)}

\item{modify_soil_depth}{truncates soil depth according to Shagguan et al. (2017)}

\item{dataset_path}{path to the 'Datasets' directory}
}
\description{
Builds a list of soil objects for a set of target point locations
}
